refactor(codegen): extract imperativeProps into plugin folder
